+++ stories/irb/FacebookPixelHuntConsent.svelte
@@ -1,0 +1,14 @@
+<script>
+    /* This Source Code Form is subject to the terms of the Mozilla Public
+    * License, v. 2.0. If a copy of the MPL was not distributed with this
+    * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+    
+    import ConsentNotice from "../../src/routes/irbs/FacebookPixelHuntConsent.svelte";
+    import IRBWindow from "../../src/routes/irbs/IRBWindow.svelte";
+    </script>
+    
+    <div style="--height: 700px;">
+    <IRBWindow>
+        <ConsentNotice />
+    </IRBWindow>
+    </div>